Adding explicit license information, since quite a large number of source files don't have the LGPL header.  This is ...
Original commit message from CVS:
Adding explicit license information, since quite a large number of source
files don't have the LGPL header.  This is something we need to fix.
GNOME has a script to run around and fix things, maybe we should snag a
copy and let it loose in a sandbox.