Closed 22145 - gstelement.c: support hotswapping of loopfunc
Original commit message from CVS:
Closed 22145 - gstelement.c: support hotswapping of loopfunc

Added a flag GST_ELEMENT_NEW_LOOPFUNC that's set when a new loopfunc is set,
which needs to be obeyed by the loopfuncs themselves.  The assumption is
that only a loopfunc or an element that knows what it's doing will ever
change the loopfunc pointer, so they can be responsible for exitting
the loopfunc itself appropriately.

The loopfunc_wrapper function has been changed to call the loopfunc
pointer itself in a loop.  If the loopfunc changes, the current loop is
supposed to exit, which causes the loopfunc_wrapper to start up the new
loopfunc.