Snapshot of work-in-progress do deal with out-of-bin elements.  Current system is very likely going to be ditched due...
Original commit message from CVS:
Snapshot of work-in-progress do deal with out-of-bin elements.  Current
system is very likely going to be ditched due to its complexity and the
fact that it doesn't work right now.  More than that, it's generating some
really odd results in my test program, which aren't readily explainable.

If you want to keep a working copy of your working copy (heh), don't update
until this is replaced with a hybrid approach.  This approach will be the
beginning of a plan generator that can construct full hybrid schedules
given hints from various places.

Hybrid means that cothreads are used, but there are chunks of the pipeline
that are actually dealt with by chaining. This can improve speed by
reducing cothread switches (which are much cheaper than thread switches,
but still far from free, about 570 cycles on my PIII), but does carry a
complexity burden.  Luckily, the structure of GStreamer allows that burden
to live entirely in create_plan.  Luck?  I think not ;-)