Fix a few irregularities, to make things closer to the Gnome standard guidelines for SGML documentation.  One change ...
Original commit message from CVS:
Fix a few irregularities, to make things closer to the Gnome standard
guidelines for SGML documentation.  One change in particular is to change
the book name for the manual from gstreamer to index; a change to the link
to the manual on the website will thus be needed when the online version
of the manual is next updated.

These changes should make it possible to view the GStreamer documentation
directly from SGML in Nautilus: anyone with a working copy of Nautilus
should try this out and report ... ;-)