Added a testsuite.  This currently only has one test, which replicates a problem just submitted to the list.
Original commit message from CVS:
Added a testsuite.  This currently only has one test, which replicates a
problem just submitted to the list.

I hope noone disagrees that we need a testsuite.  Also, hopefully we all
think that a tiny testsuite which is easy to add tests to is much better
than a huge project to create a complete testsuite, which never gets
started because its too big and scary.

I suggest that tests get created whenever someone has a spare moment, or
whenever a bug is reported (to replicate the problem, and thus be sure that
when fixed it stays fixed), and before making a change to a piece of code
to ensure that it remains working.

If we all add little tests occasionally, then we'll have a good useful
testsuite before we know it.  All you need to do is make a