Added element-level locking support, implemented for cothreads with a mutex held whenever an element is actually curr...
Original commit message from CVS:
Added element-level locking support, implemented for cothreads with a mutex
held whenever an element is actually currently running.  This should be done
with atomic variables eventually, but the best I can find is to use
<asm/spinlock.h> from the kernel, which brings in too much for my taste.

Also increased cothread stack size to 32KB and max cothread count to 64,
which fits within the default pthread 2MB stack.  It would probably be good
to do a pthread_attr_getstacksize() to verify that it's 2MB, and if
linuxthreads had a pthread_attr_setstacksize(), we could remove these
limits.