Added colorization to DEBUG and INFO.  Colors are based on categories, with the strings in _gst_category_colors[] in ...
Original commit message from CVS:
Added colorization to DEBUG and INFO.  Colors are based on categories, with
the strings in _gst_category_colors[] in gstinfo.c  See wiki node
GstDebugColors for the actual colors, or /etc/DIR_COLORS works too.

Please comment on the colorization style and ideas in the above wiki node.