Converted the DEBUG system to look like the INFO system, with a function pointer that can be replaced.  Not done with...
Original commit message from CVS:
Converted the DEBUG system to look like the INFO system, with a function
pointer that can be replaced.  Not done with everything, since it seems
that we need an ERROR system plus a FATAL error, or maybe just an ERROR
system with a severity, and depending on this the DEBUG system might again
be changed to have a single or at least fewer handlers.

Also did a bit of work fixing up the actual DEBUG and INFO commands
scattered throughout the code.  We still have over 500 of these in gst/*.c
alone...