Add the GST_ELEMENT_NO_ENTRY flag, which when set will force the scheduler to skip over an element as an option for t...
Original commit message from CVS:
Add the GST_ELEMENT_NO_ENTRY flag, which when set will force the scheduler
to skip over an element as an option for the first cothread jump.  This is
needed because libraries like mad don't ever seem to want to quit.  Thus
they ignore COTHREAD_STOPPING and can't be trusted to be the entry.