Added ability for caps_nego_failed signal to indicate that it's solved the problem, via a gboolean * argument that's ...
Original commit message from CVS:
Added ability for caps_nego_failed signal to indicate that it's solved the
problem, via a gboolean * argument that's FALSE, and can be set to TRUE if
the handler has fixed things up.

Updated the autoplugger so it now works in both the unknown and known cases
with and without other issues like a crippled (mono-only) osssink.