Fixed the media player so that it works with incsched.
Original commit message from CVS:
Fixed the media player so that it works with incsched.
Removed the thread and add the _iterate to the g_idle_loop, this makes the
GUI very cluncky but is needed because we cannot chance the state of a thread
inside the threads context yet.