Change soversion back to 0:0:0 and add use of -release flag for libtool.
Original commit message from CVS:
Change soversion back to 0:0:0 and add use of -release flag for libtool.
This means that any program linking against libgst will automatically have
the specific release of libgst encoded into it.  This enforces the fact
that (for the moment), the API/ABI is changing rapidly enough that you
can't link against 0.3.2 originally and have it still work with 0.3.3.  It
might be possible, but highly unlikely.

When we get closer to a stable API/ABI, in the 0.5.0 timeframe most likely,
we will start using soversions as recommended in the libtool docs.  Then
we have to pay more attention to forward and backwards compatiblity, or
rather, we have to *start* paying attention <g>