revert part of wingo's patch wingo, I'm not doing this to piss you off ;) but with your change we get the reverse iss...
Original commit message from CVS:
revert part of wingo's patch
wingo, I'm not doing this to piss you off ;)
but with your change we get the reverse issue : stuff in libs/gst needs
gstmarshal.[ch], which are only generated if we go in gst before libs.
So unless we just throw the egg at the chicken we'll have to settle for a
hack.
Alternatively, we could take the egg out of the chicken and put it in
a separate tarball which would then be installed ;)