Finally we're on to a proper jack setup, with a specialized bin and elements that can only go in a jack bin. I had to...
Original commit message from CVS:
Finally we're on to a proper jack setup, with a specialized bin and elements
that can only go in a jack bin. I had to fix the parser first to do this, but
to run it, the syntax is like so:

gst-launch jackbin.( filesrc ! mad ! jacksink )

But of course it's not fully functional yet. Sigh.