together with the recent commit in common, this little beauty should
Original commit message from CVS:
together with the recent commit in common, this little beauty should
- implement both --with-automake and --with-autoconf to autogen.sh
- pass them on to configure, along with good settings for aclocal and autoheader
- which then forces these tools to be used in rebuilding through running
make in the dirs (which was the whole point of this makeover)

Please test and let me know if it doesn't break other stuff.
Right, going to make the really minimal set of changes to all of our other
modules now.  I love having made a common/ dir ;)

We now return to your regularly scheduled broadcast.