Made the SWAP and PREPARE_3D_LINE macros work with gints rather than using typeof(), since typeof() is a gcc-extensio...
Original commit message from CVS:
Made the SWAP and PREPARE_3D_LINE macros work with gints rather than
using typeof(), since typeof() is a gcc-extension and does not work
with other compilers.  This is okay since every place these macros
are used, gints are passed in.  I renamed SWAP to SWAP_INT to reflect
that it is not so generic.