In the intfloat conversion elements, there were bugs when the float buffer was bigger than the int buffer (in frames)...
Original commit message from CVS:
In the intfloat conversion elements, there were bugs when the float buffer was
bigger than the int buffer (in frames). That's now fixed, some style issues are
"fixed" (subjectively), mallocs are removed from int2float's processing loop,
and the default buffer size was raised to 256 frames to give better performance in
the default case.

Wavenc was modified to not set an event handler on its sink pad. It deals
with EOS in its chain function instead.