Some prettification fixes to this interface... Make it look good, make it simple and make it more GObject'ific. This ...
Original commit message from CVS:
Some prettification fixes to this interface... Make it look good, make it simple and make it more GObject'ific. This means that it now works largely almost the same as GObject (concerning arguments that you give to the functions, and layout of the virtual functions). I've also got an OSS implementation for this, but it needs one small change before I commit that. Expect that this weekend or so.