Removed grammar.tab.h from the flex command to build lex._gst_parse_yy.c because this caused grammar.tab.h to be incl...
Original commit message from CVS:
Removed grammar.tab.h from the flex command to build lex._gst_parse_yy.c
because this caused grammar.tab.h to be included twice in the file (since
it is already included directly in parse.l via an #include statement).
Including this file twice was causing Forte to be unable to compile it.