sys/sunaudio/gstsunaudiosink.c: Do not modify the ports value. If the user has turned off the built-in speakers, then...
Original commit message from CVS:
Patch by: Brian Cameron  <brian dot cameron at sun dot com>
* sys/sunaudio/gstsunaudiosink.c: (gst_sunaudiosink_prepare):
Do not modify the ports value. If the user has turned off the
built-in speakers, then we should not reset it in the prepare
function, since this causes the built-in speakers to turn
back on anytime the user changes a track in totem, rhythmbox,
etc. (#346066).