configure.ac: Add -no-undefined to GST_ALL_LDFLAGS; without it, no shared libs will be built or installed on MingW fo...
Original commit message from CVS:
* configure.ac:
Add -no-undefined to GST_ALL_LDFLAGS; without it, no shared libs
will be built or installed on MingW for our plugins in ext/; also,
it looks like the right thing to do and we do it for all our other
plugin modules as well. Fixes #440253.