gst-libs/gst/interfaces/: Change the way these deprecated function pointers are removed so that the compiled ABI is u...
Original commit message from CVS:
* gst-libs/gst/interfaces/mixeroptions.h:
* gst-libs/gst/interfaces/mixertrack.h:
Change the way these deprecated function pointers are removed
so that the compiled ABI is unconditionally smaller. This
sets in stone an ABI break that actually occurred when the
things were deprecated in 0.10.14, which seems to be the best
fix as the only known users are oss-mixer and sunaudio-mixer in
gst-plugins-good.
Fixes: #513018