po: avoid conflicts of local *.po files with files in git
Make it so that filenames and line numbers are only stored in the *.pot file (which is not in git), but not in the
individual *.po files. This information is hardly useful for translators in our case, and it should avoid the constant
conflicts of local *.po files with the ones in git which are caused by the source files changing and the line numbers
being updated.

This commit is likely to cause one last merge conflict for you, which you can work around with "git checkout po/*.po"
before merging or pulling. After that there should (hopefully) not be any more local modifications of these files.