xdgmime: make xdg typefinder more conservative
The whole raison d'etre of this typefinder is to help avoid false
positives when used in combination with our main typefinder in
-base. Its task is not really to typefind audio/video files, even
less so given that it detects *MIME* types, not GStreamer media
types. Therefore, if this typefinder detects an audio or video
type, don't even suggest it - our own typefinders are hopefully
better at this and detect the right type of type on top of that,
and even if they're not we really want to know about it and fix
it. Fixes #575157.