po: avoid conflicts of local *.po files with files in git
Make it so that filenames and line numbers are only stored in the *.pot file
(which is not in git), but not in the individual *.po files. This information
is hardly useful for translators in our case, and it should avoid the constant
conflicts of local *.po files with the ones in git which are caused by the
source files changing and the line numbers being updated. This commit might
cause one last merge conflict for you, which you can work around with
"git checkout po/*.po" before merging or pulling. After that there should
(hopefully) not be any more local modifications of these files (unless
someone committed additions or changes to translated strings and the
*.po files haven't been updated yet, that is).