Remove the search_func/user_data parameters to GstIndex::get_assoc_entry
They're not used in any Index implementation and we can't use them
in C# either. This should work without problems because according to
C calling conventions the caller is responsible for preparing
the function arguments and cleaning up later. This only means that
func and user_data will contain invalid data now.