resindvd: Remove extra new-segment in rsnstreamselector. Send discont.
All the input pads for the resin stream selectors have the same source,
and receive/send the same segment info, so there's no need to send a
segment when switching and activating a different pad. Removing it makes
on-the-fly audio track switching work properly.

After switching audio track, mark the next buffer discont.