win32: change how win32/common/config.h is updated
Generate win32/common/config.h-new directly from config.h.in,
using shell variables in configure and some hard-coded information.
Change top-level makefile so that 'make win32-update' copies the
generated file to win32/common/config.h, which we keep in source
control. It's kept in source control so that the git tree is
buildable from VS.

This change is similar to the ones applied a while ago to GStreamer
core, gst-plugins-base and gst-plugins-good and prevents configure
from changing files in source control. The generated config.h
should be ok, but needs testing.