vcdsrc: change VCD URI handler to the DVD scheme
The current URI parsing code doesn't allow setting the "device"
from which the VCD should be read.

Use the same structure as the DVD URI handling instead, so that
devices can be passed in the URI, as well as track number.

Up the rank of the VCD plugin so that it can be auto-plugged and
used by Totem.

https://bugzilla.gnome.org/show_bug.cgi?id=340986