configure: Remove -Wcast-align
Apparently gcc warns that GstMiniObject is not castable to
GstEvent/Message/Buffer due to them containing 64bit variables, even
though ARM hackers claim that those only need 4byte alignment. And as
long as gcc behaves that way, this warning is not very useful.

So we'll remove the warning until this problem is fixed.

https://bugzilla.gnome.org/show_bug.cgi?id=615698