videomixer: Don't return caps in get_caps() that will be rejected
This commit basically puts _get_caps() in sync with accept_caps().

If we don't have a master pad OR the master pad caps aren't negotiated
then we just return the downstream allowed caps.

If we have a master pad with negotiated caps, we return those caps
with a free range of width/height/framerate