Use AG_GST_CHECK_GST_PLUGINS_{GOOD,UGLY,BAD,FFMPEG} macros from common
to find plugin directories for the various modules our unit tests depend
on in an uninstalled environment. This makes sure these plugins are found
even when distchecking (which happens from a subdirectory, which means
that the currently used trick to find the uninstalled directories of
these modules doesn't work in that case).