camerabin: Refactor tests
The tests were done in 2 steps, first there was a suite
that generated the files (while checking that camerabin
was operating correctly). Then there was a second suite
that was run to check that all files were playable with
playbin2. Those second tests were not being run because
they were checking if camerabin was initialized, and it
never was as those tests didn't use a 'setup' function.

This commit refactors the tests by removing this second
suite and merging its validation with the first suite's
functions.